fix: RandomInt32Generator.cs should have int generic arg type :)

+++ src/Stl/Generators/RandomInt32Generator.cs
@@ -5,7 +5,7 @@ using System.Security.Cryptography;
 namespace Stl.Generators
 {
     // Thread-safe!
-    public sealed class RandomInt32Generator : Generator<long>
+    public sealed class RandomInt32Generator : Generator<int>
     {
         private readonly byte[] _buffer = new byte[sizeof(int)];
         private readonly RandomNumberGenerator _rng;
@@ -13,7 +13,7 @@ namespace Stl.Generators
         public RandomInt32Generator(RandomNumberGenerator? rng = null)
             => _rng = rng ?? RandomNumberGenerator.Create();
 
-        public override long Next()
+        public override int Next()
         {
             lock (_rng) {
                 _rng!.GetBytes(_buffer);
